# contracts/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Template API",
    "version": "v1"
  },
  "paths": {
    "/api/v1/auth/capabilities": {
      "get": {
        "tags": [
          "Template.Api"
        ],
        "operationId": "GetAuthCapabilities",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfAuthCapabilitiesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "Template.Api"
        ],
        "operationId": "GetAuthSession",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfAuthSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "tags": [
          "Template.Api"
        ],
        "operationId": "GetAuthCsrf",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfAuthCsrfResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Template.Api"
        ],
        "operationId": "Logout",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Request token returned by GET /api/v1/auth/csrf.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfAuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          }
        ]
      }
    },
...
    "/api/local-auth/scenario": {
      "post": {
        "tags": [
          "local-only"
        ],
        "operationId": "CreateLocalAutomationScenario",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Request token returned by GET /api/v1/auth/csrf.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocalAutomationScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfLocalAutomationScenarioResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-local-only": true
      },
      "delete": {
        "tags": [
          "local-only"
        ],
        "operationId": "DeleteLocalAutomationScenario",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Request token returned by GET /api/v1/auth/csrf.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfLocalAutomationCleanupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": [ ]
          }
        ],
        "x-local-only": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseOfAuthCapabilitiesResponse": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuthCapabilitiesResponse"
          }
        }
      },
...
      "ProblemDetails": {
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "traceId"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "enum": [
              "invalid_request",
              "validation_failed",
              "unauthorized",
              "forbidden",
              "not_found",
              "method_not_allowed",
              "internal_error",
              "antiforgery_failed",
              "local_auth_invalid_credentials",
              "local_auth_user_required",
              "local_auth_disabled",
              "local_auth_user_exists",
              "rate_limited",
              "invalid_return_url",
              "external_provider_not_configured",
              "already_authenticated",
              "external_auth_failed",
              "external_email_required",
              "external_email_unverified",
              "external_identity_conflict",
              "external_email_conflict",
              "oauth_flow_context_changed",
              "invalid_cursor",
              "external_connection_required",
              "external_connection_not_found",
              "account_session_not_found",
              "current_session_cannot_be_revoked",
              "concurrency_conflict",
              "organization_ownership_transfer_required",
              "organization_not_found",
              "organization_permission_denied",
              "organization_name_conflict",
              "organization_slug_conflict",
              "last_organization_required",
              "organization_confirmation_mismatch",
              "member_not_found",
              "target_user_not_found",
              "member_already_exists",
              "member_role_unchanged",
              "role_assignment_forbidden",
              "member_domain_acknowledgement_required",
              "team_not_found",
              "team_permission_denied",
              "team_name_conflict",
              "team_name_unchanged",
              "team_member_not_found",
              "team_member_already_exists",
              "invitation_not_found",
              "invitation_permission_denied",
              "invitation_already_exists",
              "invitation_recipient_already_member",
              "invitation_team_invalid",
              "invitation_domain_restricted",
              "invitation_recipient_mismatch",
              "invitation_email_verification_required",
              "invitation_expired",
              "invitation_not_pending",
              "invitation_membership_conflict",
              "invitation_limit_reached"
            ],
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "email": {
            "maxLength": 254,
            "type": [
              "null",
              "string"
            ],
            "description": "Target email for member_domain_acknowledgement_required.",
            "format": "email"
          },
          "emailDomain": {
            "maxLength": 253,
            "pattern": "^(?=.{1,253}$)(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$",
            "type": [
              "null",
              "string"
            ],
            "description": "Normalized target domain for member_domain_acknowledgement_required."
          },
          "allowedEmailDomains": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "maxLength": 253,
              "pattern": "^(?=.{1,253}$)(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$",
              "type": "string"
            },
            "description": "Ordered allowed domains for member_domain_acknowledgement_required."
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "description": "Secure HttpOnly same-origin session cookie.",
        "name": "__Host-template.session",
        "in": "cookie"
      }
    }
  },
...
}